refactor: Response, Request 분리 - controller(response), domain(request)
