+++ src/main/java/site/travellaboratory/be/article/domain/_schedule/request/ScheduleEtcRequest.java
@@ -1,0 +1,7 @@
+package site.travellaboratory.be.article.domain._schedule.request;
+
+public record ScheduleEtcRequest(
+    String placeName
+) {
+
+}
