+++ src/main/java/site/travellaboratory/be/article/domain/request/ArticleUpdateRequest.java
@@ -1,4 +1,4 @@
-package site.travellaboratory.be.article.presentation.response.writer;
+package site.travellaboratory.be.article.domain.request;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -6,7 +6,7 @@ import java.util.List;
 
 public record ArticleUpdateRequest(
         String title,
-        List<LocationDto> locations,
+        List<LocationRequest> locations,
         LocalDate startAt,
         LocalDate endAt,
         String expense,
