--- src/main/java/site/travellaboratory/be/article/presentation/response/_schedule/writer/ScheduleEtcRequest.java
@@ -1,7 +1,0 @@
-package site.travellaboratory.be.article.presentation.response._schedule.writer;
-
-public record ScheduleEtcRequest(
-    String placeName
-) {
-
-}
